Size Block with u32 rather than usize
Setting the width and height with usize is overly cautious and is not in
congruence with the use of u32 in other sections of the library.
Specifically, the Element types et al. are set up with u32 types. The
conversion to usize adds unnessecary friction.

+++ src/block.rs
@@ -8,20 +8,20 @@ pub(crate) trait DrawBlock {
 
 pub(crate) struct Block {
     /// Width in pixels.
-    pub(crate) width: usize,
+    pub(crate) width: u32,
     /// Height in pixels.
-    pub(crate) height: usize,
+    pub(crate) height: u32,
     /// Row-major pixel buffer.
     pub(crate) buf: Vec<Pixel>, // TODO: Invariant: buf.len() == width * height
 }
 
 impl Block {
     /// Creates a new [`Block`].
-    pub(crate) fn new(width: usize, height: usize, background: Pixel) -> Self {
+    pub(crate) fn new(width: u32, height: u32, background: Pixel) -> Self {
         Self {
             width,
             height,
-            buf: vec![background; width * height],
+            buf: vec![background; width as usize * height as usize],
         }
     }
 
@@ -31,7 +31,7 @@ impl Block {
     ///
     /// If `self.width` is 0, the function will panic.
     pub(crate) fn rows(&self) -> Rows {
-        self.buf.chunks_exact(self.width)
+        self.buf.chunks_exact(self.width as usize)
     }
 
     /// Returns an iterator over mutable rows in this [`Block`].
@@ -40,7 +40,7 @@ impl Block {
     ///
     /// If `self.width` is 0, the function will panic.
     pub(crate) fn rows_mut(&mut self) -> RowsMut {
-        self.buf.chunks_exact_mut(self.width)
+        self.buf.chunks_exact_mut(self.width as usize)
     }
 
     // TODO: Improve this weirdly worded doc comment.
@@ -54,22 +54,22 @@ impl Block {
     ///
     /// In case `start_x` is greater than the width of the `Block` that is painted onto
     /// (`self.width`), this function will panic.
-    pub(crate) fn paint(&mut self, other: Self, start_x: usize, start_y: usize) {
+    pub(crate) fn paint(&mut self, other: Self, start_x: u32, start_y: u32) {
         if other.width == 0 || self.width == 0 {
             return;
         }
         assert!(start_x <= self.width);
         assert!(start_y <= self.height);
-        let delta_x = usize::min(other.width, self.width - start_x);
-        let delta_y = usize::min(other.height, self.height - start_y);
+        let delta_x = u32::min(other.width, self.width - start_x);
+        let delta_y = u32::min(other.height, self.height - start_y);
         let end_x = start_x + delta_x;
         for (row, other_row) in self
             .rows_mut()
-            .skip(start_y)
-            .take(delta_y)
+            .skip(start_y as usize)
+            .take(delta_y as usize)
             .zip(other.rows())
         {
-            row[start_x..end_x].copy_from_slice(&other_row[..delta_x])
+            row[start_x as usize..end_x as usize].copy_from_slice(&other_row[..delta_x as usize])
         }
     }
 
@@ -85,7 +85,7 @@ impl Block {
             "pixel buffer is not large enough"
         );
         for (y, row) in self.rows().enumerate() {
-            let idx = (y * self.width) * PIXEL_SIZE;
+            let idx = y * self.width as usize * PIXEL_SIZE;
             // TODO: See if we can get rid of this iter(). Perhaps through feature(slice_flatten)?
             // TODO: Where should the .copied() go, ideally?
             let row_bytes: Vec<_> = row.iter().copied().flatten().collect();
